[KV offload][1b/N] rename offloading to kv_offload (#25191)
