+++ tests/v1/kv_offload/test_worker.py
@@ -1,7 +1,7 @@
 # SPDX-License-Identifier: Apache-2.0
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
-from vllm.v1.offloading.abstract import LoadStoreSpec
-from vllm.v1.offloading.worker.worker import (OffloadingHandler,
+from vllm.v1.kv_offload.abstract import LoadStoreSpec
+from vllm.v1.kv_offload.worker.worker import (OffloadingHandler,
                                               OffloadingWorker, TransferResult,
                                               TransferSpec)
