+++ vllm/v1/kv_offload/worker/worker.py
@@ -3,7 +3,7 @@
 from abc import ABC, abstractmethod
 
 from vllm.logger import init_logger
-from vllm.v1.offloading.abstract import LoadStoreSpec
+from vllm.v1.kv_offload.abstract import LoadStoreSpec
 
 # a single transfer spec (src_blocks_spec, dst_blocks_spec)
 TransferSpec = tuple[LoadStoreSpec, LoadStoreSpec]
